Moved cassandra schema into it's own file
Since we often want to reference it, better that the schema not be
hidden away in cassandra/helpers.go
